refactor: runner
